+++ src/components/LanguageSwitcher.tsx
@@ -1,7 +1,7 @@
-import { useTranslation } from 'react-i18next';
-import Box from '@mui/material/Box';
 import ToggleButton from '@mui/material/ToggleButton';
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
+import { useTranslation } from 'react-i18next';
+import { Box }  from '@mui/material';
 
 const LanguageSwitcher = () => {
   const { i18n } = useTranslation();
@@ -15,7 +15,7 @@ const LanguageSwitcher = () => {
   return (
     <Box
       display="flex"
-      alignItems="center"  
+      alignItems="center"
       paddingRight={2}
       marginRight={2}
       borderRight={1}
@@ -27,8 +27,12 @@ const LanguageSwitcher = () => {
         exclusive
         onChange={handleLanguageSwitch}
       >
-        <ToggleButton value="fr" style={{color: 'white', backgroundColor: i18n.language === 'fr' ? 'grey' : 'transparent'}}>FR</ToggleButton>
-        <ToggleButton value="en" style={{color: 'white', backgroundColor: i18n.language === 'en' ? 'grey' : 'transparent'}}>EN</ToggleButton>
+        <ToggleButton value="fr" style={{ backgroundColor: i18n.language === 'fr' ? 'grey' : 'transparent' }}>
+          🇫🇷
+        </ToggleButton>
+        <ToggleButton value="en" style={{ backgroundColor: i18n.language === 'en' ? 'grey' : 'transparent' }}>
+          🇬🇧
+        </ToggleButton>
       </ToggleButtonGroup>
     </Box>
   );
